+++ db/migrate/20140305191733_remove_screen_token.rb
@@ -1,0 +1,5 @@
+class RemoveScreenToken < ActiveRecord::Migration
+  def change
+    remove_column :users, :screen_token, :string
+  end
+end