+++ db/migrate/20140305192954_remove_service_account_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveServiceAccountFromUsers < ActiveRecord::Migration
+  def change
+    remove_column :users, :service_account, :boolean, default: false, null: false
+  end
+end