+++ test/controllers/projects_controller_test.rb
@@ -151,21 +151,6 @@ class ProjectsControllerTest < ActionController::TestCase
     assert_template 'index'
   end
 
-  test "should get index for api user using service account" do
-    login(users(:service_account))
-    get :index, api_token: 'screen_token', screen_token: users(:valid).screen_token, format: 'json'
-    assert_not_nil assigns(:projects)
-
-    projects = JSON.parse(@response.body)
-    assert projects.first.keys.include?('id')
-    assert projects.first.keys.include?('name')
-    assert projects.first.keys.include?('color')
-    assert projects.first.keys.include?('favorited')
-    assert projects.first.keys.include?('tags')
-
-    assert_response :success
-  end
-
   test "should get new" do
     get :new
     assert_response :success