+++ test/test_helper.rb
@@ -33,7 +33,6 @@ class ActionDispatch::IntegrationTest
     user.update_column :status, user_template.status
     user.update_column :deleted, user_template.deleted?
     user.update_column :system_admin, user_template.system_admin?
-    user.update_column :service_account, user_template.service_account?
     post_via_redirect 'users/login', user: { email: email, password: password }
     user
   end