+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 - Use of Ruby 2.1.1 is now recommended
 - Enabled turbolinks
+- Removed Screen dependencies and service accounts
 - **Gem Changes**
   - Updated to rails 4.0.3
   - Updated to contour 2.4.0.beta3